fix(logic): fix program logic and some entity

+++ src/main/java/com/chinese_dictation/model/entity/UserProgress.java
@@ -36,7 +36,7 @@ public class UserProgress {
     private Integer currentSentenceIndex;
 
     private Integer totalAttempts;
-    private Integer totalTimeSpentSeconds;
+    private Long totalTimeSpentSeconds;
 
     @CreatedDate
     private LocalDateTime createdAt;
